requirements.txt
requests==2.32.*
load_dotenv==0.1.*
python-telegram-bot==13